src/requirements.txt
django==3.2.5
psycopg2-binary==2.8.5
djangorestframework==3.12.4
